requirements.txt
dj-database-url
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
gunicorn
Pillow
psycopg[binary]
uvicorn
uvicorn-worker
whitenoise[brotli]